Add IMS:LTI support
Added support for lti calls.
Update requirements file.

- To use LTI the LMS manager needs to add a LTI-Admin tool on each
course to grant permission on an experiment. This tool is an open
window to the labmanager LMS-admin-view.

- The Course admin needs to add a second tool to launch the labmanager
experiments section.
When an Instructor or Learner access the tool, a list with all the
allowed experiments will be displayed and a selection of the experiment
will present it.